refactored scaling code
