+++ public/js/calendar/calendar.js
@@ -1,6 +1,7 @@
 import { formatAMPM } from "../utils/formatAMPM.js";
 import { getEventInfo } from "./getEventInfo.js";
 import { getScheduledMeetingsByDate } from "./getMeetingsForCalendar.js";
+import { scaleUpElement, resetScaleElement } from "../utils/scaleElements.js";
 
 document.addEventListener("DOMContentLoaded", function () {
   const calendarContainer = document.getElementById("calendar");
@@ -238,19 +239,6 @@ document.addEventListener("DOMContentLoaded", function () {
 
     calendarContainer.innerHTML = "";
     calendarContainer.appendChild(calendarTable);
-  }
-
-  function scaleUpElement(element) {
-    if (element.tagName === "TD") {
-      element.style.transform = "scale(1.2)";
-    } else {
-      element.style.transform = "scale(1.1)";
-    }
-    element.style.transition = "transform 0.1s ease";
-  }
-  function resetScaleElement(element) {
-    element.style.transform = "scale(1)";
-    element.style.transition = "transform 0.1s ease";
   }
 
   function highlightDate(element) {
